auth/requirements.txt
click==8.3.0
Flask==2.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2==2.9.11
PyJWT==2.3.0
Werkzeug==2.0.3
